refactor: lizeWindowをtsに移行
